requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
imageio==2.36.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==3.0.2
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
threadpoolctl==3.5.0
tifffile==2024.9.20
webcolors==24.11.1
Werkzeug==3.1.3
